Task_5/requirements.txt
streamlit
faiss-cpu
sentence-transformers
ollama
numpy
pickle